% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append-table-name.R
\name{append_table_name}
\alias{append_table_name}
\title{function to append aliases in code
called without arguments}
\usage{
append_table_name()
}
\description{
function to append aliases in code
called without arguments
}
\examples{
demo_sql("rooms"); append_table_name()
open_sql("inst/rooms.sql")
}
