% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append-table-name.R
\name{append_text}
\alias{append_text}
\title{Line by line function to swap out text}
\usage{
append_text(code, table)
}
\description{
Line by line function to swap out text
}
\examples{
append_text(code = "select max(case when 123 between x and y then 1 else 0 end) as w from sys.db.visit", table = "visit")
}
