% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompts.R
\name{ask_and_wait}
\alias{ask_and_wait}
\title{print request to the console}
\usage{
ask_and_wait(text)
}
\description{
print request to the console
}
\examples{
ask_and_wait(text = "press enter to continue")
}
