% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open-sql.R
\name{demo_sql}
\alias{demo_sql}
\title{open demo sql file in temp directory
there are some files in the 'ref' folder but you can point this at any sql file}
\usage{
demo_sql(demo = c("rooms", "reservations", "ortho-query"))
}
\arguments{
\item{demo}{string of sample sql files to use}
}
\description{
open demo sql file in temp directory
there are some files in the 'ref' folder but you can point this at any sql file
}
\examples{
demo_sql()
demo_sql("rooms")
}
