% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace-aliases.R
\name{find_alias_replacements}
\alias{find_alias_replacements}
\title{build list of alias patterns}
\usage{
find_alias_replacements(x)
}
\description{
build list of alias patterns
}
\examples{
x <- find_statement(statement = "highlight a from statement")$selection[[1]]$text
find_alias_replacements(x)
}
