% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbtify.R
\name{find_sources}
\alias{find_sources}
\title{Identify the sources from a sources yml file}
\usage{
find_sources(path = "etl/sources.yml")
}
\description{
Identify the sources from a sources yml file
}
\examples{
find_sources()
}
