% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompts.R
\name{find_statement}
\alias{find_statement}
\title{find from or select statements}
\usage{
find_statement(statement)
}
\description{
find from or select statements
}
\examples{
find_statement(statement = "the 'from' statements you want to use")
}
