% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbtify.R
\name{jinjify}
\alias{jinjify}
\title{convert from statement to source & ref
called without arguments}
\usage{
jinjify(sources_path = "etl/sources.yml")
}
\description{
convert from statement to source & ref
called without arguments
}
\examples{
jinjify()
}
