% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyword-patterns.R
\name{keyword_patterns}
\alias{keyword_patterns}
\title{Title}
\usage{
keyword_patterns(table)
}
\arguments{
\item{table}{string of table name to find and append to keywords}
}
\description{
Title
}
\examples{
keyword_patterns(table = "mtcars")
}
