% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open-sql.R
\name{open_sql_as_temp}
\alias{open_sql_as_temp}
\title{open sql file in temp directory}
\usage{
open_sql_as_temp(file)
}
\arguments{
\item{file}{string of sql file location}
}
\description{
open sql file in temp directory
}
\examples{
open_sql_as_temp(file = "")
}
