% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace-aliases.R
\name{replace_aliases}
\alias{replace_aliases}
\title{function to replace aliases in code
called without arguments}
\usage{
replace_aliases()
}
\description{
function to replace aliases in code
called without arguments
}
\examples{
test_code(); replace_aliases()
}
