% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace-tab-with-space.R
\name{replace_tab_with_space}
\alias{replace_tab_with_space}
\title{called without arguments}
\usage{
replace_tab_with_space()
}
\description{
called without arguments
}
\examples{
demo_sql(); replace_tab_with_space()
}
