% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbtify.R
\name{source_replacements}
\alias{source_replacements}
\title{Generate regex replacements for tables found in sources.yml}
\usage{
source_replacements(path)
}
\description{
Generate regex replacements for tables found in sources.yml
}
\examples{
source_replacements()
}
