% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlize.R
\name{sqlize}
\alias{sqlize}
\title{sqlize}
\usage{
sqlize(df, table, na_string = "NULL", file = "sqlizeR", save = TRUE)
}
\arguments{
\item{df}{data.frame}

\item{table}{string containing name of the SQL table}

\item{na_string}{string to replace NA values with}

\item{file}{string containing file prefix to save .sql file if save=TRUE}

\item{save}{boolean to save result to file (default), otherwise print to screen}
}
\description{
This function creates SQL insert statements for a dataframe
}
\examples{
data(PlantGrowth)
sqlize(PlantGrowth, table="plant_growth", save=FALSE)
}
